Use a u32 rather than u64 because we have to support power pc.
And who will ever need more than 32 bits, anyway?

+++ src/cmd.rs
@@ -193,7 +193,7 @@ impl server::Output for PrintlnOutput {
         println!("{}", output);
     }
 
-    fn provide_id(&self) -> u64 {
+    fn provide_id(&self) -> u32 {
         0
     }
